requirements.txt
Aiogram~=3.4.1
redis~=5.0.1
gspread==6.0.0
oauth2client==4.1.3
aiohttp==3.9.5
pytz
apscheduler
Pillow
sentry-sdk
